+++ src/notification-service/src/presentation.rs
@@ -1,8 +1,7 @@
-use common::{currency::domain::Currency, language::domain::Language};
+use common::language::domain::Language;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub struct NotificationPresentationPreferences {
     pub language: Language,
-    pub currency: Currency,
     pub prohibited_content_consent: bool,
 }